Add additional context for readability and optimization
